+++ jeap-spring-boot-monitoring-starter-it-webflux/src/test/java/ch/admin/bit/jeap/monitor/TestConfig.java
@@ -1,7 +1,0 @@
-package ch.admin.bit.jeap.monitor;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestConfig {
-}