+++ jeap-spring-boot-security-starter-it-webflux/src/test/java/ch/admin/bit/jeap/security/it/client/webflux/TestConfig.java
@@ -1,7 +1,0 @@
-package ch.admin.bit.jeap.security.it.client.webflux;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestConfig {
-}