+++ jeap-spring-boot-security-starter-test/src/main/java/ch/admin/bit/jeap/security/test/resource/configuration/DisableJeapOAuth2SecurityConfiguration.java
@@ -1,12 +1,10 @@
 package ch.admin.bit.jeap.security.test.resource.configuration;
 
 import ch.admin.bit.jeap.security.resource.configuration.MvcSecurityConfiguration;
-import ch.admin.bit.jeap.security.resource.configuration.WebFluxSecurityConfiguration;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ImportAutoConfiguration(exclude = {MvcSecurityConfiguration.class, WebFluxSecurityConfiguration.class
-})
+@ImportAutoConfiguration(exclude = MvcSecurityConfiguration.class)
 public class DisableJeapOAuth2SecurityConfiguration {
 }