+++ jeap-spring-boot-security-starter-test/src/main/java/ch/admin/bit/jeap/security/test/resource/configuration/PermitAllWebSecurityConfiguration.java
@@ -7,10 +7,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.web.server.ServerHttpSecurity;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.server.SecurityWebFilterChain;
-import org.springframework.security.web.server.util.matcher.ServerWebExchangeMatchers;
 
 @AutoConfiguration
 @SuppressWarnings("Convert2MethodRef")
@@ -27,21 +24,6 @@ public class PermitAllWebSecurityConfiguration {
                     securityMatcher("/**").
                     csrf(csrf -> csrf.disable()).
                     authorizeHttpRequests(authorizeHttpRequests -> authorizeHttpRequests.anyRequest().permitAll());
-            return http.build();
-        }
-    }
-
-    @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.REACTIVE)
-    @Configuration
-    public static class ReactivePermitAllWebSecurityConfiguration {
-        @Bean
-        @Order(Ordered.HIGHEST_PRECEDENCE + 5)
-        // Overrule other web security configurations, but still allow room for overriding.
-        public SecurityWebFilterChain permitAllSecurityWebFilterChain(ServerHttpSecurity http) {
-            http.
-                    securityMatcher(ServerWebExchangeMatchers.pathMatchers("/**")).
-                    csrf(csrf -> csrf.disable()).
-                    authorizeExchange(exchanges -> exchanges.anyExchange().permitAll());
             return http.build();
         }
     }