+++ jeap-spring-boot-swagger-webflux-starter/src/test/java/ch/admin/bit/jeap/swagger/webflux/it/TestConfiguration.java
@@ -1,7 +1,0 @@
-package ch.admin.bit.jeap.swagger.webflux.it;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-class TestConfiguration {
-}