+++ jeap-spring-boot-web-config-starter-it-webflux/src/main/java/ch/admin/bit/jeap/web/configuration/webflux/TestApp.java
@@ -1,7 +1,0 @@
-package ch.admin.bit.jeap.web.configuration.webflux;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestApp {
-}