+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [20.0.0] - 2026-03-11
+
+### Changed
+
+-  Breaking Change
+  - **Removed**
+    - Support for reactive/webflux
+    - Support removed from monitoring, tracing, swagger, security web-config starters
+
 ## [19.16.0] - 2026-03-10
 
 ### Changed